duui-ner-spacy/src/main/docker/requirements.txt
spacy==3.7.5
dkpro-cassis==0.9.1
uvicorn==0.30.5
fastapi==0.112.0
